+++ Sources/MigrationKit/SchemaSnapshotProvider.swift
@@ -1,0 +1,4 @@
+public protocol SchemaSnapshotProvider {
+    associatedtype Snapshot
+    func generateCanonicalSnapshot() throws -> Snapshot
+}